Updated channel data type in LimeConfig_t
Switched the data type of 'channel' from float to int in LimeConfig
structure to ensure channel numbers are represented as integers,
aligning with typical usage and avoiding potential truncation errors.
This change improves configuration accuracy and safeguards against
invalid fractional channel values.

+++ src/limedriver.h
@@ -29,7 +29,7 @@
 struct LimeConfig_t {
 
   float srate;
-  float channel;
+  int channel;
   float frq;
   float frq_set;
   float RX_LPF;
